+++ migrations/Version20210819125057.php
@@ -1,0 +1,33 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DoctrineMigrations;
+
+use Doctrine\DBAL\Schema\Schema;
+use Doctrine\Migrations\AbstractMigration;
+
+/**
+ * Auto-generated Migration: Please modify to your needs!
+ */
+final class Version20210819125057 extends AbstractMigration
+{
+    public function getDescription(): string
+    {
+        return '';
+    }
+
+    public function up(Schema $schema): void
+    {
+        // this up() migration is auto-generated, please modify it to your needs
+        $this->addSql('ALTER TABLE case_documents ADD soft_deleted TINYINT(1) DEFAULT \'0\' NOT NULL');
+        $this->addSql('ALTER TABLE document DROP soft_deleted');
+    }
+
+    public function down(Schema $schema): void
+    {
+        // this down() migration is auto-generated, please modify it to your needs
+        $this->addSql('ALTER TABLE case_documents DROP soft_deleted');
+        $this->addSql('ALTER TABLE document ADD soft_deleted TINYINT(1) DEFAULT \'0\' NOT NULL');
+    }
+}