+++ src/Logging/EntityListener/AbstractEntityListener.php
@@ -8,6 +8,7 @@ use App\Logging\ItkDevGetFunctionNotFoundException;
 use App\Logging\ItkDevLoggingException;
 use App\Logging\LoggableEntityInterface;
 use Doctrine\ORM\Event\LifecycleEventArgs;
+use Doctrine\ORM\PersistentCollection;
 use Symfony\Component\Security\Core\Security;
 use Symfony\Component\Uid\UuidV4;
 
@@ -31,34 +32,52 @@ abstract class AbstractEntityListener
         $em = $args->getEntityManager();
 
         $object = $args->getObject();
+
+        $isRemoveAction = false;
+
         $changeArray = $em->getUnitOfWork()->getEntityChangeSet($object);
 
+        // If empty then it must be remove/delete action
+        if (empty($changeArray)) {
+            $changeArray = $em->getUnitOfWork()->getOriginalEntityData($object);
+            $isRemoveAction = true;
+        }
+
+        // Array with change data
         $dataArray = [];
 
         foreach ($changeArray as $key => $value) {
-            if (!array_key_exists(1, $value)) {
-                $message = 'Value array does not contain new value.';
-                throw new ItkDevLoggingException($message);
+            $changedValue = null;
+
+            if ($isRemoveAction) {
+                $changedValue = $value;
+            } else {
+                if (!array_key_exists(1, $value)) {
+                    $message = 'Value array does not contain new value.';
+                    throw new ItkDevLoggingException($message);
+                }
+
+                $changedValue = $value[1];
             }
 
-            $changedValue = $value[1];
-
-            // We do not log properties with value null
-            // todo determine if we should
+            // In case a nullable property is edited to null we must log this
             if (null === $changedValue) {
+                $dataArray[$key] = '';
                 continue;
             }
 
+            // ID is already present as entity_id no need to add it twice
             if ($changedValue instanceof UuidV4) {
-                $dataArray[$key] = $changedValue->__toString();
                 continue;
             }
 
+            // Handle DateTime(s)
             if ($changedValue instanceof \DateTime) {
                 $dataArray[$key] = $changedValue->format('d-m-Y H:i:s');
                 continue;
             }
 
+            // Handle loggable entities
             if ($changedValue instanceof LoggableEntityInterface) {
                 $dataArray[$key] = $this->handleLoggableEntities($changedValue);
                 continue;
@@ -69,21 +88,34 @@ abstract class AbstractEntityListener
                 continue;
             }
 
+            // Logging is done on case level, no need to log case 'twice'
+            if ($changedValue instanceof CaseEntity) {
+                continue;
+            }
+
+            if ($changedValue instanceof PersistentCollection) {
+                continue;
+            }
+
+            // Property was not handled
             $message = sprintf('Unhandled property %s of type %s.', $key, get_class($changedValue));
             throw new ItkDevLoggingException($message);
         }
 
+        // Create log entry
         $logEntry = new LogEntry();
 
+        // Set values on log entry
         $logEntry->setCaseID($case->getId());
         $logEntry->setEntity(get_class($object));
         $logEntry->setEntityID($object->getId());
         $logEntry->setAction($action);
 
         $user = $this->security->getUser();
-
-        if (null !== $user) {
+        if ($user) {
             $logEntry->setUser($user->getUsername());
+        } else {
+            $logEntry->setUser('Fixtures');
         }
 
         $logEntry->setData($dataArray);