+++ CHANGELOG.md
@@ -35,6 +35,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 MunicipalityListener, BoardListener and their tests.
 - [TVIST1-256](https://jira.itkdev.dk/browse/TVIST1-256): Added case number and CaseManager service.
 - [TVIST1-264](https://jira.itkdev.dk/browse/TVIST1-264): Added fixtures.
+- [TVIST1-184](https://jira.itkdev.dk/browse/TVIST1-184): Added ability to upload documents and showing
+  uploaded documents in document index.
+- [TVIST1-185](https://jira.itkdev.dk/browse/TVIST1-185): Added ability to 'soft' delete documents.
+- [TVIST1-186](https://jira.itkdev.dk/browse/TVIST1-186): Added ability to copy documents to other cases.
+- [TVIST1-126](https://jira.itkdev.dk/browse/TVIST1-126): Adding workflows for cases.
+- [TVIST1-227](https://jira.itkdev.dk/browse/TVIST1-227): Added ability to create, edit and remove notes.
 
 ### Changed
 - [NSEK-138](https://jira.itkdev.dk/browse/NSEK-138): Updating Cypress to 6.5.0 so the same version is used project wide.